Teaching_Image_Generator/requirements.txt
streamlit==1.14.0
requests==2.26.0
python-dotenv==0.19.2